Дорога - новые модели. Иконки башен и дороги, Выстрел, след снаряда и попадание для башен и замка.

+++ Assets/Scripts/Game/State/Root/GameplayState.cs
@@ -15,7 +15,7 @@ namespace Game.State.Root
         public int MapId { get; set; }
         public int CurrentWave { get; set; }
         public int GlobalEntityId { get; set; }
-        public int GameSpeed { get; set; } = 1;
+        public float GameSpeed { get; set; } = 1f;
         //public int PreviousGameSpeed { get; set; }
         public int Progress { get; set; } //Текущий прогресс игры, растет от убийства мобов, при = наполнении обнуляем и увеличиваем ProgressLevel на 1
         public int ProgressLevel { get; set; } //Влияет на коэффициент роста Progress в обратно-пропорциональном порядке
